service/requirements_service.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.114.1
h11==0.14.0
idna==3.8
numpy==2.1.1
pandas==2.2.2
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
pyarrow==17.0.0
pydantic==2.9.1
pydantic_core==2.23.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
